refactor(app): #373 Phase 1-5 use-layout-resize.ts に panel resize を切り出し
App.tsx 1656 -> 1535 行 (-121)。Claude Code パネル / サイドバーの drag
リサイズ・CSS 変数同期・dblclick リセットを新規 hook
src/renderer/src/lib/hooks/use-layout-resize.ts に集約。

移管:
- 定数: MIN_PANEL / MAX_PANEL / DEFAULT_PANEL / MIN_SIDEBAR /
  MAX_SIDEBAR / DEFAULT_SIDEBAR
- ref: resizeDragRef / sidebarResizeDragRef
- effect: settings 復元時に CSS 変数 (--claude-code-width /
  --shell-sidebar-w) を初期化する 2 本
- handler: onClaudePanelResizeStart / onSidebarResizeStart /
  onSidebarResizeDouble (旧 handleResizeStart / handleSidebarResizeStart /
  handleSidebarResizeDouble)

Phase 1-1 〜 1-4 と異なり opts を一切取らない (settings の読み書きのみで
完結する純粋な責務のため)。useSettingsValue / useSettingsActions を
hook 内で直接呼ぶ流儀。

drag 中の React 再レンダリング回避 (CSS 変数を documentElement.style に
直接 setProperty) と、mouseup gating での updateSettings 呼び出し
(debounce ではない) も完全保持。body / handle 要素への is-resizing /
is-dragging クラス付与・剥がしも現行どおり。

検証:
- npm run typecheck: green
- cargo check --manifest-path src-tauri/Cargo.toml: green
- cargo clippy --no-deps: 既存ベースライン 15 件のまま (新規警告ゼロ)

累計削減 (Phase 1-1 〜 1-5): -1292 行 (#373 全体目標 -1995, 進捗 65%)。

Refs #373
